[OC-558], move icon component into the folder navbar
