+++ ui/main/src/app/app.module.ts
@@ -18,7 +18,7 @@ import {AppRoutingModule} from './app-routing.module';
 import {AppComponent} from './app.component';
 import {NavbarComponent} from './components/navbar/navbar.component';
 import {LoginComponent} from "./components/login/login.component";
-import {IconComponent} from './components/icon/icon.component';
+import {IconComponent} from './components/navbar/icon/icon.component';
 import {TranslateModule} from "@ngx-translate/core";
 import {translateConfig} from "./translate.config";
 import {FontAwesomeModule} from '@fortawesome/angular-fontawesome';
