+++ ui/main/src/app/components/navbar/icon/icon.component.html
